refactor: separate cli from code
